pyproject.toml
[tool.poetry]
name = "meow"
version = "0.1.0"
description = "A simple, fast and powerful enterprise email finder tool was built by antx, which based on pocx."
authors = ["antx-code <wkaifeng2007@163.com>"]

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple"
default = true

[tool.poetry.dependencies]
python = "^3.8"
pocx = "^0.2.2"
PyYAML = "^6.0"
pyhunter = "^1.7"
uvloop = "^0.16.0"
typer = "^0.6.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
